tests: Move TestLogBackend to runner
